m03: fix: Capture enemy frigate objective not triggering
Obsoletes 2f480a53b5e34ad08d2602955f6cc92025429381

This bug was lurking probably since the original release:
Ships rendered at LOD 0 (the highest level of detail) did not trigger the RenderedShips function.
The bitmask used by that function always had 0 (false) for LOD 0.
This was never an issue in the original game because a ship had to fill
almost the entirety of the screen to get up to LOD 0;
So you would always have gone through at least LOD 1 before, and the
events always triggered correctly.
However, with our recent Full Detail tweak, ALL ships are at LOD 0. ALWAYS.

So I fixed the mask, and mission objectives trigger correctly again :)

I also cleaned up redundant checks as RenderedShips checks lower LODs as well.

I also, also, re-encoded Mission03.kas as UTF-8. I'm not sure whether
that'll break anything, but I couldn't find a single encoding that
worked for the whole file. ISO 8859-2 seems to encompass all the special
characters in the translated dialogues, but leaves entire rows of ÄÄÄÄ
in the comments. Those render as ──── in CP 437 which is the sanest I
have found in the different encodings available to me, but this encoding
breaks all the translated dialogue.
So I manually fixed that and saved it all to UTF-8.